Fix warning from sysinfo.dwNumberOfProcessors on Windows

+++ src/client/CPUCount.cpp
@@ -33,11 +33,10 @@
 
 // Number of CPU cores
 int CPUCount() {
-
 #ifdef _WIN64
     SYSTEM_INFO sysinfo;
     GetSystemInfo(&sysinfo);
-    return sysinfo.dwNumberOfProcessors;
+    return (int) sysinfo.dwNumberOfProcessors;
 #elif __APPLE__
     int count = 0;
     size_t len = sizeof(count);
